test (BonusNumberDtoTest): 검증 기능 이동에 따른 테스트 코드 수정
